# swagger.json
{
	"components": {
...
		"schemas": {
			"UserI": {
				"properties": {
					"_id": {
						"type": "string",
						"description": "Mongo Document Id"
					},
					"username": {
						"type": "string",
						"description": "Username"
					},
					"email": {
						"type": "string",
						"description": "User email"
					},
					"userId": {
						"type": "string",
						"description": "Business User Id"
					},
					"password": {
						"type": "string",
						"description": "Password encrypted"
					},
					"firstName": {
						"type": "string",
						"description": "user first name"
					},
					"lastName": {
						"type": "string",
						"description": "User last name"
					},
					"active": {
						"type": "boolean",
						"description": "Is the user active?"
					},
					"externalUser": {
						"type": "boolean",
						"description": "External User. If true is external user that has participation in OpenChallenges"
					},
					"areaVisible": {
						"items": {
							"$ref": "#/components/schemas/AreaI"
						},
						"type": "array",
						"description": "Area that is visible for user"
					},
					"updated": {
						"type": "string",
						"format": "date-time",
						"description": "Last user update"
					}
				},
				"required": [
					"username",
					"email",
					"userId",
					"password",
					"firstName",
					"lastName",
					"active",
					"externalUser"
				],
				"type": "object",
				"additionalProperties": true
			},
			"CompanyI": {
				"properties": {
					"_id": {},
					"companyId": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"CNPJ": {
						"type": "string"
					},
					"committe": {
						"items": {
							"$ref": "#/components/schemas/UserI"
						},
						"type": "array"
					}
				},
				"required": [
					"companyId"
				],
				"type": "object",
				"additionalProperties": true
			},
			"AreaI": {
				"properties": {
					"name": {
						"type": "string"
					},
					"company": {
						"$ref": "#/components/schemas/CompanyI"
					}
				},
				"required": [
					"name",
					"company"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"GroupValidatorI": {
				"properties": {
					"groupValidatorId": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					},
					"area": {
						"$ref": "#/components/schemas/AreaI"
					}
				},
				"required": [
					"groupValidatorId",
					"name",
					"created",
					"area"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"CommitteI": {
				"properties": {
					"active": {
						"type": "boolean"
					},
					"leader": {
						"$ref": "#/components/schemas/UserI"
					},
					"general": {
						"items": {
							"$ref": "#/components/schemas/UserI"
						},
						"type": "array"
					}
				},
				"required": [
					"active"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"IntegrantI": {
				"properties": {
					"user": {
						"$ref": "#/components/schemas/UserI",
						"description": "Validator User"
					},
					"integrantId": {
						"type": "string",
						"description": "committe integrant Id"
					},
					"active": {
						"type": "boolean",
						"description": "is active?"
					},
					"created": {
						"type": "string",
						"format": "date-time",
						"description": "When was created?"
					},
					"lastChangePosition": {
						"type": "string",
						"format": "date-time",
						"description": "Date of the last change of position in the committe. \nFor example from LEADER to GENERAL or GENERAL to leader"
					},
					"finished": {
						"type": "string",
						"format": "date-time",
						"description": "Date end"
					},
					"groupValidator": {
						"$ref": "#/components/schemas/GroupValidatorI",
						"description": "committe integrant can be part of group validator\n. Just Committe integrants can be validators"
					},
					"role": {
						"type": "string",
						"description": "What is the role? Committe General or Committe Leader"
					}
				},
				"required": [
					"user",
					"integrantId",
					"active",
					"created",
					"lastChangePosition",
					"role"
				],
				"type": "object",
				"additionalProperties": true
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "ideation_backend",
		"version": "1.0.0",
		"description": "ideation backend",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Clara Idea"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/committe/leader": {
			"post": {
				"operationId": "NewLeader",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommitteI"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "string"
							}
						}
					}
				}
			}
		},
		"/committe": {
			"post": {
				"operationId": "NewCommitte",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
...
		"/integrant/leader/{integrantId}": {
			"post": {
				"operationId": "NewLeader",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IntegrantI"
								}
							}
						}
					}
				},
				"description": "Endpoint for add or reeplace a leader of committe.\nThe new leader have to be a `general commite` user first",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "integrantId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/integrant/general": {
			"post": {
				"operationId": "NewIntegrant",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IntegrantI"
								}
							}
						}
					}
				},
				"description": "Endpoint used for add a new integrant to committe. \nUser could be a new integrant or old integrant with active=false",
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "user that will be converted from generator to commite (GENERAL)",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "string"
							}
						}
					}
				}
			}
		},
...
		"/integrant/general/{integrantId}": {
			"delete": {
				"operationId": "DeleteLeader",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "integrantId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}